QSDK:CP ath12k:wifi: Fix queuing peer delete command during SSR mode1 falling back to mode0
Usage of the ATH12K_FLAG_RECOVERY is to indicate that SSR is in
progress with mode 0. This flag will be set when we receive SSR
notifier call back during firmware crash. In case of mode0, this
flag will be set to all the radios (asserted and partner chips)
This flag is also used to stop queueing peer delete to firmware.

When mode 1 fallbacks to mode 0 and in the process when host sends
peer delete, the WMI command gets queued to CE but peer unmap is not
received which is expected because q6 crashed. When the timeout
happens warning is printed which is not expected.

When mode 1 SSR fallbacks to mode 0, host will trigger fw_hang
command and will reset UMAC recovery or mode 1 flags from partner
links. But host sets recovery mode as mode 0, it doesn't set mode 0
related flags. Before partner ab crashes and firmware exit is
received, a peer delete gets trigged in host for which host sends
peer delete WMI cmd. This command will be posted to CE ring because
ATH12K_FLAG_CRASH_FLUSH is not yet set for the partner chips. This
makes host to wait for firmware response and triggers warn on.

To fix the issue set mode 0 ATH12K_FLAG_RECOVERY flag when SSR mode
falls back to mode0 and host will clear ATH12K_FLAG_RECOVERY
reconfig completion.


Patch-dependency: 707-b-ath12k-add-support-for-MLO-SSR-Mode0.patch
Patch-work: None

Change-Id: I4d54dd0e834712526008b9e7d32b8b8fd9c6a00c
Signed-off-by: Harish Rachakonda <quic_rachakon@quicinc.com>