QSDK:CP: mac80211.sh: Enable syslog for MLD single instance in hostapd
By default, hostapd is enabled with INFO debug logs, but these logs are
redirected to logread only if the syslog option is enabled.

This option is already enabled for the global instance.
Enable the syslog option for the MLD single instance of hostapd to
ensure that INFO debug logs are properly redirected to logread.

Change-Id: Ife443ea781515523e4714c3ee83575098c7bf3ca
Signed-off-by: Govindaraj Saminathan <quic_gsaminat@quicinc.com>