UPSTREAM:DP: wifi: ath12k: Adjusted RCU Lock in PPDU stats
Refactor the handling of rcu_lock for ar during PPDU stats collection.
Move the spinlock_bh outside the RCU lock to prevent RCU stalls
during stats collection, as ar is not an RCU-protected variable and
unlocking the RCU at the end of stats collection is unnecessary.

Patch-dependency: None
patch-work: None

Change-Id: Ic1938c146dafdb63b5919f9ae7c46f8ff21f1337
Signed-off-by: Gautham Kumar Senthilkumaran <quic_gauthamk@quicinc.com>