QSDK:PROTOCOL:wifi: mac80211: Fix the puncturing bitmap in mesh vdev
When the NL80211_CMD_JOIN_MESH command is received, the puncturing
bitmap is included as part of chan_def, but it is not updated in link_conf->eht_puncturing. The ath12k driver uses link->conf->eht_puncturing to apply the puncturing bitmap during vdev start, which results in the puncturing bitmap not being applied
in vdev.

To fix the issue, update the chan_def puncturing bitmap in link_conf->eht_puncturing.

Patch-dependency:
638-ath12k-Send-Puncturing-pattern-of-peer-mesh-during-a.patch

Patch-work: None

Change-Id: I4d5f147c41581d60980ca52aa8f360572dab7bd2
Signed-off-by: Ramasamy Kaliappan <quic_rkaliapp@quicinc.com>