QSDK:CP: Report sta add failure to mac80211 during recovery
When SSR mode1 is triggered in one of the device, during this
time if any legacy station add is given from mac80211, driver
is returning success without actually adding the station to
driver during recovery.

Because of this, when mac80211 sends station removal after SSR
mode 1 recovery, driver is unable to find the station and triggers
the warn on.

Add changes to send station add failure to mac80211 during recovery
and overwrite any of the removal errors as 0 to avoid warn_on.

Patch-Dependency: 901-wifi-ath12k-Peer-clean-up-during-subsystem-recovery.patch
Patch-work: None

Change-Id: Ib3ea40ef858054a86b8df57cf3171bd0082c830f
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>