QSDK:CP : wifi: ath12k: Accept vendor wifi param for legacy interface
Fix NSS adjustment for legacy clients when the upper layer sends a WiFi
param for a legacy interface. Ensure the set WiFi param subroutine
handles valid legacy interfaces by checking for the absence of the MLO
link attribute and retrieving the radio context from the deflink of an
interface, thereby correcting the NSS adjustment for a valid radio.

Currently, peer addition and deletion logs are printed with ath12k_info,
which floods the console. Move these logs to the debug level.

Change-Id: I6c8473d714112beea6e71c28b05f1c55595bdc21
Signed-off-by: Sidhanta Sahu <quic_sidhanta@quicinc.com>