QSDK:CP: mac80211.sh: Fix the input validation issues in mlo link addition
Currently, if user gives the link addition command to add
a link which is already present in the interface, that link
will be added in the MLD which is incorrect and fix this issue
with validation check.

Also, change the mbssid value to 3 and add the mbssid_group_size
as 4 for 6GHz link addition.

Limitation:
We can't add a MLO link op top of SLO bring-up with /etc/config/
wireless file as wifi mlo_add_link expects the mld config in the
/etc/config/wireless to add a link.

Fixes: 9455dbeaba (mac80211.sh: Add support to handle mlo links addition)

Change-Id: I3bf4c27367aa28ce82ae931379202662b7ce8ceb
Signed-off-by: Manish Dharanenthiran <quic_mdharane@quicinc.com>
Signed-off-by: Maharaja Kennadyrajan <quic_mkenna@quicinc.com>