QSDK:PLATFORM mac80211: Prevent wdev registration if wiphy is not registered.
During rmmod, wdev list is cleared,and the wiphy is unregistered.
However, wdevs are sometimes added after the wiphy has been unregistered,leading to a WARN_ON.
Add a check to ensure that the wiphy is still registered before proceeding with the registration of wireless devices. If the wiphy is not registered,trigger a WARN_ON to identify the process attempting to add wireless devices after the wiphy has been unregistered.

Change-Id: Ia2df857e105c0ea9b22756d4fb29f8c688b3ad88
Signed-off-by: Narendhar Reddy <quic_nare@quicinc.com>