QSDK:CP: wifi: ath12k: Handle num of peers in arvif for SSR mode 1
During SSR mode1, arvif->num_peers are not cleared properly
because of which there is a mismatch between arvif->num_peers and
ar->num_stations. When wifi command is invoked after this, while
clean up ar->num_station is going to negative value because of
which none of the clients are able to connect back.

Hence, handle num of peers in arvif for SSR mode 1 case.

Fixes: I6368bd9ec23 ("UPSTREAM:CP:wifi:ath12k: Add peer delete all support WMI cmd")

Change-Id: Ia30693a6f626ba2cfbb3bcaae42373b529af639d
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>