UPSTREAM:CP ath12k: Restrict removing monitor interface during SSR recovery
When the firmware is in the recovery process, there are instances
where mac80211 might attempt to remove the monitor interface and host
sends WMI command to FW to stop monitor vap and waits for FW response
as FW is in recovery response is not received. This result in warnings
being printed and unexpected behavior.

Fix this by stop queueing monitor vdev stop command to firmware for device under going recovery until firmware powers up again.

Patch-dependency: 707-b-ath12k-add-support-for-MLO-SSR-Mode0.patch
Patch-work: None

Change-Id: I46c7d4fd6dde6b9a4dc12f716e024ec0621974c9
Signed-off-by: Harish Rachakonda <quic_rachakon@quicinc.com>