QSDK:CP:wifi:ath12k: Unassign link STA when there is failure in link removal during recovery
During Mode-1 recovery when there is a genuine link removal request from
hostapd after partner peer cleanup and before recovery reconfig, Firmware
won't be having the peer entries this will lead to a failure in
ath12k_station_unauthorize(). Because of this ath12k_mac_station_remove()
will never gets invoked and the arsta contents will stay intact but no
error will be reported to mac80211. Because of this mac80211 will assume
it to be a successful link removal and remove the corresponding link_sta
entry from ieee80211_sta.

This results in a stale entry in ath12k_sta. Fix this by freeing the memory
of that corresponding ath12k_link_sta when there is a failure in
ath12k_station_unauthorize() during recovery.

Fix this by freeing the link sta contents when there is such failure.

Patch-dependency:
	970-005-wifi-ath12k-Handle-link-sta-delete-for-AP-STA-mode-s.patch

Change-Id: I512a1dfd7d6560fefab899cd263e4b5b8a3a1801
Signed-off-by: Avula Sri Charan <quic_asrichar@quicinc.com>
Signed-off-by: Gautham Kumar Senthilkumaran <quic_gauthamk@quicinc.com>