QSDK:CP: scripts: Handle duplicate link addition with wifi command
1) Assign a name to the WiFi interface configuration
   (e.g., config wifi-iface 'default_radio0_band1') in the
   wireless configuration.
2) Remove the link dynamically using hostapd_cli.
3) Add the removed link dynamically using the wifi command.

This process results in the link being added, causing the wireless
configuration to have two wifi-ifaces under the same MLD in the
same radio. However the expectation is, the link should not be added
as the wireless config already has a link in that radio under that MLD.

In this same scenario, instead of adding the removed link back, if a
link is added in the radio where there is already a link configured
under that MLD, then it creates two vdevs in same radio under an MLD
which is an issue. Then during wifi down, when the links are
removed, the firmware asserts due to this two vdev scenario.

Fix this issue by removing the usage of the "@" character with
link addition, ensuring that any valid name given/not-given to the
wifi-iface works properly.

Fixes: 89f1bfde50b3 ("QSDK:CP: mac80211.sh: Fix the input validation issues in mlo link addition")

Change-Id: I0e6bc2d7a77a97f2d0ae89ce9ff9530e5c72440a
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>