QSDK:PROTOCOL: wifi: ath12k: Update CCA counter after each beacon Tx
Currently, during CCA, the CSA counter offset is used to update
CCA counter to FW. When CSA is not active but CCA event happens,
the counter doesn't update CCA count, causing failure of the event.

Fix this issue by updating beacon with new countdown for every
beacon tx completion event and not using the csa counter.

Earlier, this implementation was removed since for EMA, all EMA
beacons are updated in a single instance and counter update for every
Tx event will lead to incorrect countdown sequence over the air.
Now, as MBSSID grouping is enabled by default, we do not encounter
this issue.

Also, currently additional checks are enabled in FW which restricts
the usage of CSA counter offset for CCA, and hence moving back to the
older design. The CCA countdown sequence issue with EMA in the older
design will still be exposed and can be fixed only if FW offload is
available for CCA.

Since this implementation was present previously, deleting the patch
that removed this change.

Feature: HE BSS Color

Change-Id: Iacdb4abff3c2b1f2d96a829540ca400d0f74eb79
Signed-off-by: Mohan Kumar G <quic_mkumarg@quicinc.com>