UPSTREAM:CP: wifi: ath12k: perform MLD STA mac address validation
Currently when one legacy STA is associated with AP with mac address X,
now when an multi link STA tries to associate that has same mld mac address
X as of legacy STA. In case when mld mac address is different from the
MLD STA link's mac address, during cleanup if MLd STA is removed for any
reason the data for legacy STA is also remove causing corruption on firmware
and vice versa. Due to this firmware crash is seen.

To resolve this issue add a check for MLD mac address of STA during peer
create.

Patch-dependency: None
Patch-work: None

Fixes: If051a4194d1 ("ath12k: perform address validation during ML peer creation")

Change-Id: Iacbdc78b7b29aaf089d9735d4094a41e6aedb176
Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>