QSDK:CP:wifi:ath12k: handle dp peer setup failure
Even when peer is not found during dp peer setup, sta state machine
is trying to process the sta data in  which leads to a null pointer
crash as the sta entry is not available in driver.

Handle failure case for dp peer setup and add debug logs in case
of peer link removal.

Fixes: I810f7e29ba7 ("wifi: ath12k: allow primary link selection")

Change-Id: I936b463c627907d98f61a1b58f506d4340b20258
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>