UPSTREAM:PLATFORM:wifi:ath12k: Fix retriggering of UMAC reset when already one is in progresss
When DP UMAC reset interrupt ATH12K_UMAC_RESET_INIT_TARGET_RECOVERY_SYNC_USING_UMAC
is received twice as a result trigger done bit to FW is set twice which causes
FW crash as already UMAC reset is in progress when trigger done bit is set for
the second time. Skip setting of trigger bit when recovery is already in progress.

Add WARN_ON when UMAC reset init interrupt is received twice from
FW as this is not expected.

Change-Id: I1a3b1d6afa29f201de9bf1d86d6d5273b920cc9e
Signed-off-by: Sowmiya Sree Elavalagan <quic_ssreeela@quicinc.com>