UPSTREAM:CP:wifi:ath12k: Skip assigning channel context if radio is not active
In case of multiple countries mentioned in the wireless config file the
regulatory rules will be updated for every different country. But  if
in between of these events if there's radio_start request from mac80211
the radio will comply with the regulatory rules at that point of time
which might sometimes lead to bringing down a radio.

Post this if there's a new regdb available where some of the previously
disabled radio's are valid here, mac80211 will try to create and start
vdev on a radio that is not even active with respect to the driver.

To avoid this add an additional check in ath12k_get_ar_by_ctx() to check
if the radio is active or not and return NULL if it's notactive. This
will ensure that any request from mac80211 on in active radio will be
discarded.

Patch-dependency:
	652-02-ath12k-Update-hw-generic-mac-ops-handling.patch

Change-Id: I0a3a41bfe57b4221de61b88c31419d1f75f7741d
Signed-off-by: Avula Sri Charan <quic_asrichar@quicinc.com>