UPSTREAM:CP wifi:ath12k Restrict the acquisition of ieee80211_bss_conf for bridge vdev
In the function ath12k_peer_assoc_h_eht, restrict the acquisition of
bss_conf for bridge vdev, as it is an internally created vdev for
diagonally associating peers.

Due to this failure, EHT capabilities are not being filled for the
bridge peer, causing firmware to assert when sending WMI_PEER_ASSOC_CMD.

Fix this by restricting the acquisition of bss_conf for bridge vdev
and setting MC15 to disable for the bridge peer.

Patch-dependency:
	936-UPSTREAM-PROTOCOL-wifi-ath12k-Send-MCS15-support-to-.patch
	961-01-wifi-ath12k-n-link-mlo-bridge-vdev-basic-handling.patch

Patch-work: none

Fixes: I71a8416f885:(UPSTREAM:PROTOCOL:wifi: mac80211: Send mcs15 support to FW)

Change-Id: Ic308418159c35cb7e69f9bf34b0e98f4915b86c5
Signed-off-by: Vignesh C <quic_vignc@quicinc.com>