QSDK:PROTOCOL:wifi:ath12k: Fix validation of AWGN bitmap
When AWGN interference occurs, an interference bitmap is received through
WMI. Based on this bitmap, either a channel change or bandwidth reduction
is triggered. Before taking any action, the received interference bitmap
undergoes validation. However, there is an issue with this validation
process. For instance, if the current operating bandwidth is 320 MHz, the
received bitmap is validated against the pattern 0xFF00, which only
considers interference in the secondary 160 MHz. If interference affects
all subchannels except the primary 20 MHz, the received bitmap will be
0xFFFE, and this bitmap is incorrectly deemed invalid.

To address this issue, the validation process should also consider the
primary segment subchannels when validating the interference bitmap.

Change-Id: I308fb91b81748cf112a7c0ee185d94d2795b5b25
Signed-off-by: Amith A <quic_amitajit@quicinc.com>