UPSTREAM:CP: wifi: ath12k: Avoid beacon update during SSR
When SSR mode0 is triggered, for one of the device the reset
work is queued after 40 seconds approximately. During this time
user space application detects that there is no stats received
and tries to update the beacon.

When driver receives this notification, it sees like arvif is already
started and throws warning but in actual, the firmware is in assert
and these flags should have been cleared. But, due to the delay it is
not cleared.

Add changes to avoid channel context/beacon update during ssr mode 0.

Change-Id: Id44882cef8786dd7d488a566368d6b08aa3094ee
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>