UPSTREAM:PROTOCOL: ath12k: Fix Max Number of Simultaneous Links
Currently, the Maximum Number of Simultaneous Links in MLD
is set to the value received from the firmware,
which is 4 in all cases.

However, the Maximum Number of Simultaneous Links should
be (N-1) (number of radios - 1).

Fix this issue by setting the Maximum Number of Simultaneous
Links to the value of (N-1).

Fixes: I56c721364 ("wifi: ath12k: propagate ML capabilities")

Patch-dependency: None

Patch-work: [PATCH] wifi: ath12k: Fix Max Number of Simultaneous Links

Change-Id: I899219cdd82595faefc4ea302bbf883a5fa2c321
Signed-off-by: Yuvarani V <quic_yuvarani@quicinc.com>