QSDK:DP:wifi:ath12k: optimize ring full condition check
In the dp tx path, the tcl ring full condition with volatile
access to tail ptr is done for every packet.
This is not optimal.

The following are known behavior for data path  rings.
a) tp always follows hp.
b) ring full condition is not a very frequent event
(i.e it does not happen for every packet).

Since tp always follows hp, if hp is not equal to last cached tp,
it means that ring is not full.

Since this condition is sufficient to put a new entry into the ring,
remove the volatile access of tp for every packet and fetch the real tp
only if ring full condition is hit with last cached tp.

This change gives ~1200Mbps improvement in 240MHz TBTC case and ~5% improvement
of CPU in UDP DL in RDP433.

This change gives ~490Mbps improvement in 160MHz TBTC case and ~10% improvement
of CPU in UDP DL in RDP433.

Patch-depdendency: None
Patch-work: None

Change-Id: Ic45edfa219cd78801e06a0ae8ebd4b4f3891274c
Signed-off-by: Hari Chandrakanthan <quic_haric@quicinc.com>