UPSTREAM:PROTOCOL: Replace CAC timeout delayed work with HR timer
Prior initiating communication in a DFS channel, there should be a
monitoring of RADAR in that channel for a minimum of 600 seconds if it
is a Weather RADAR channel and 60 seconds for other DFS channels. This
Channel Availability Check(CAC) is currently implemented by scheduling
a work item for execution with a delay equal to an appropriate timeout.
But this work item is observed to take more delay than specified
(4-5 seconds in regular DFS channels and 25-30 seconds in Weather RADAR
channels). Even though this delay is expected in case of delayed work
queue as there is no guarantee that the work will be scheduled exactly
after the specified delay, a delay of more than 20 seconds is too much
for the AP to be in non-operational state.

Starting from 4.8 Kernel, Timer wheel implementation has changed for
better design. This has resulted in loss of accuracy
(worst case - up to 12.5%). Following is the commit that made the change
https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/commit/?id=500462a9de657
This inherent delay is resulting in the delay observed in CAC timeout.

Therefore, to have a better precision for the CAC timeout, use High
Resolution(HR) timer instead of delayed work queue.

Patch-depdendency: None

Patch-work: Replace CAC timeout delayed work with HR timer

Feature: DFS

Change-Id: I5251b7bc33e61f6b1b58227cabf1487116d2b2c5
Signed-off-by: Amith A <quic_amitajit@quicinc.com>