UPSTREAM:DP:wifi:ath12k:Remove WAR for vp number issue in splitphy
A Workaround (WAR) has been implemented to address an issue where
the VP number might be incorrectly mapped to the wrong lmac_id,
causing a traffic halt in the split PHY MLO case.

This issue is addressed in the change ID I10a6895c97, hence removing
the WAR in the wifi-classifier path.

Added MAC Debug for the WiFi classifier path to improve debuggability.

Patch-dependency: 970-QSDK-ath12k-Reuse-vp-profiles-in-case-of-splitphy

Change-Id: Ib881530e626bc7088f2081a4cce79b5f53732346
Signed-off-by: Girish Kumar M P <quic_gmp@quicinc.com>