QSDK:CP: wifi-script: Enhance multi link configuration handling
mlo_remove_link enhancement:

In mlo_remove_link we are removing wireless config for that link.
Due to this information like wds, encryption, etc. are lost and if user tries
to add link back to interface then these information are not added back.

Fix this by not deleting link from wireless config. Disable the link when
link removal is done.

mlo_add_link/mlo_remove_link enhancement:

Scripts rely on mac address to identify link_id, channel and
ru_punct_bitmap of a link.

Better way is to use frequency range to identify the parameters and remove
dependency from mac address.

mlo_remove_link enhancement:
For example 2 3 link mld are present, now user removes one link from one
interface then all the wireless config of that device is removed. This is
because the validation if check is not proper.

Fix this by also checking the respective mld of that iface before disabling
the link.

Change-Id: I7ef95d6468ad1978aaa6480094db526a5abde306
Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>