QSDK:PLATFORM:CP mac80211:Prevent interface addition if wiphy is not registered.
During rmmod, wdev list is cleared,and the wiphy is unregistered.
if hostapd crashes and restarts during rmmod it is sending
NL command to add interface, which is leading to interface
addition even though wiphy is unregistered.

Add a check to prevent interface addition if the wiphy is
unregistered.

Error:
Unable to handle kernel access to user memory outside uaccess routines at virtual address 0000000000000068
[17616.479294] Mem abort info:
[17616.489987]   ESR = 0x0000000096000005

[17616.978765] Call trace:
[17616.985879]  cfg80211_get_drvinfo+0x20/0x17c [cfg80211]
[17616.988138]  ethtool_get_drvinfo+0x50/0x244
[17616.993345]  dev_ethtool+0x54c/0x1620
[17616.997512]  dev_ioctl+0x3e8/0x47c
[17617.001330]  sock_do_ioctl+0xb4/0xfc

Change-Id: I3161a3153914b8380d2ecf0079de3cb9f5c8396d
Signed-off-by: Narendhar Reddy <quic_nare@quicinc.com>