QSDK:DP:wifi:ath12k: Fix RCU stall due to register write when Q6 is down
During mode 0 recovery when rootPD is down and coredump collection is in progresss,
mac remove interface would be called for removing vlan interface created
for WDS repeater which calls a register write to unset Tx PPE VP entry but this stalls
the CPU as Q6 is down.
Registers should not be accessed when Q6 is down and hence  skipping the register
write if ATH12K_FLAG_RECOVERY flag is set.

Change-Id: Ic0fcec110216dd7317ebd7f834990b8f364ed944
Signed-off-by: Sowmiya Sree Elavalagan <quic_ssreeela@quicinc.com>