QSDK:DP: boost_performance.sh: Disable Linux RPS
During 3-link DL KPI measurement, received packet are not evenly
distributed to cores in STA side which causes a low throughput.
This is because of linux RPS was't disabled since FSE based RPS
core selection is in use. Hence, disable the Linux RPS in script
to avoid multiple time core selection of received packet.

Fixes: I2a5ba8c (clean redundant commands, optimize ath12k settings)

Change-Id: Ic2b6b787df4cb00f4eaa3a9efb7f3d63408ba9dc
Signed-off-by: Thiraviyam Mariyappan <quic_tmariyap@quicinc.com>