QSDK:CP: hostapd.uc: ucode: set sec_chan offset based on chanwidth
In WDS case, RE AP will be brought up with the same channel configuration as of RE STA. If RE DUT is configured with EHT80
and when RE STA connects with Root AP on chan width 20, even
though wpa_s passes the sec_channel as 0 for 20MHz bandwidth, in function iface_freq_info(), its assigned as NULL to auto calculate.

This makes the sec_channel_offset as 1 on uc_wpa_freq_info() and makes iface start call with wrong freq info. As an result of this, RE AP
wont come up.

To fix this, overwrite sec_channel based on chan_width if the channel params are passed from wpa_s.

Change-Id: Ia50822909d6d973a0d48944fd9915d63e178bc47
Signed-off-by: Vignesh C <quic_vignc@quicinc.com>