QSDK:CP: wifi: ath12k: Use hw_idx as primary link
Currently link id from the iw dev output, which is a non-variable
value for each band, is set in the wireless config file. With the
new OpenWrt ucode changes, the link id value in the iw dev output
becomes dynamic, affecting consistency after reboot.

Use the hw_idx value from the iw phy00 info, which remains
constant after reboot as the primary link in the wireless config
file and set the respective link id for that hw_idx in the driver.

Patch-work: None
Patch-dependency: None

Change-Id: Ie2452a6b5f03243172b1f484a1d06a38a5583357
Signed-off-by: Maharaja Kennadyrajan <quic_mkenna@quicinc.com>