UPSTREAM:CP:wifi: Fix random address corruption in failure case
WMI control path stats allocate dynamic memory for the stats, update as
list and free it after printing the stats. However the stats are freed
during failure case such as memory allocation failure, ar is not found
for the pdev idx. In such failure scenarios, parent function also has a
mechanism to free the memory by looping through all the list and freeing
the memory as well as deleting the list

In one such failure case, when host is not able to get ar from pdev id,
memory is freed but list added to the stats is not deleted. Since the
stats list not empty, during the next free random memory is deleted.

Patch-work:
https://lore.kernel.org/linux-wireless/202501220916.o70hNSkA-lkp@intel.com/T/#m8c23b6f423d37666d98d4e2cb0d7d857dbcef632

Change-Id: I74d0e50339437c15286bb1b098248b8314fcc97d
Signed-off-by: Ramya Gnanasekar <quic_rgnanase@quicinc.com>