QSDK:CP: wifi: Update the proper msduq in metadata
Sometimes host sends out the metadata with invalid msduq
(0x3F) to NSS; whereas NSS expects it to be 0 if the
peer msduq (peer_id + msduq) is invalid (0xFFFF).

So the expected metadata here in non-mlo case should be
0xAA800000 instead of 0xAAC0003F.

--------------------------------------------------------------------
| tag  | mlo_key_valid| sawf_valid| reserved| mlo key| peer_id| msduq |
|8 bits|   1 bit      |  1 bit    | 1 bit   | 5 bits | 10 bit | 6 bits|
----------------------------------------------------------------------

Earlier NSS used to bypass the metadata which was sent from
host and do the classification. But due to a recent change
in ecm classifier, ecm expects the metadata to be sent with
proper data.
This change helps in updating the metadata with proper data.

Patch-dependency: None
Patch-work: None
Fixes: None

Change-Id: I38a364d8d1e0cdf01d44d21ed0f6c8f9b4e4ef8b
Signed-off-by: Adwait Nayak <quic_adwaitn@quicinc.com>