UPSTREAM:CP: wifi: ath12k: Enhance wmi ctrl stats
Currently, wmi ctrl path stats frees all the items in the wmi list while
processing tlvs of each tag. With the enablement of periodic stats, it
is evident that the wmi list cannot be used to print all tags.

On-demand tag is expected when an user requests for wmi ctrl path stat
for a specific tag via debugFs of ar `wmi_ctrl_stats`. If the stat is
not read, then the previously requested data is not freed until a new
input is provided by the user. User can read the stats using the above
debugFs file to print the stats and at the same time it frees the data
from the wmi list.

However, in case of periodic wmi ctrl path stats, on every periodic
event, it frees the existing data from the wmi list and add the new one.
In case a user requests for an on demand stat when periodic stat is
enabled, it frees all the items from the list due to periodic event.

Add support to print periodic and on-demand wmi ctrl path stats
simultaneously.

Change List:
        - Add new wmi debug list - period_wmi_list for storing the
          periodic event
        - Use existing wmi_list for on-demand stats
        - Add tagid under struct wmi_ctrl_path_stats_to print the
          specific stats from the un-freed list
        - Re-factor the print function of all wmi ctrl stats to work
          with this new change

Patch-dependency: 982-04-QSDK-CP-wifi-ath12k-Add-support-for-QCA-Energy-Service.patch
Patch-work: None

Change-Id: Ieae44c989d205c5b7a58ff50125b06f69e0886be
Signed-off-by: Sidhanta Sahu <quic_sidhanta@quicinc.com>