UPSTREAM:PLATFORM: Prevent parallel qmi sequence in case of consecutive FW crash
As per design qmi.num_radios need to be filled with INVALID_RADIO during
core alloc and during recovery to make sure qmi seqeunce is sequential
across SOC's.

Curently for Mode-1 recovery this is being taken care in
ath12k_core_mlo_hw_queues_stop() which will set the qmi.num_radios
of asserted ab to INVALID_RADIO.

And for Mode-0 recovery driver will check if partner chip is already
in recovery and if it's not  in recovery then the qmi.num_radios of
partner chip will be set to INVALID_RADIO and fw hang command will
be sent for that partner chip.

But in case of conseqcutive FW crashes this will not be updated in some
SOC's qmi.num_radios which will result in parallel QMI sequence.Fix this
by updating the num_radios across SOC's in such corner cases.

Change-Id: I638c6c44a9a6ca03414332db8ed70e2fa2f975d3
Signed-off-by: Avula Sri Charan <quic_asrichar@quicinc.com>