QSDK:CP:PLATFORM: Fix DS traffic stall after MODE0 recovery
Traffic stall issue is happening in DS mode after MODE0 recovery
is completed. This is due to not mapping the correct VP profile
with VP table entry for the VP number assigned to the VIF. In DS,
for every link creation, DS attach for the link with the VP number,
and when the link goes down, detach the link for the VP number happens.

But in MODE0 recovery, the link up sequence is called without the
link down sequence. Due to that, the DS detach link will not be done.
And during the link up sequence, the DS attach link happens again for
the same link and VP number, choosing the wrong VP profile with
VP table entry. This creates an invalid AST entry issue in TQM that
triggers packet drops in FW and leads to a traffic stall issue.

To avoid this problem, call the DS detach link function in
ath12k_core_halt() during MODE0 recovery to avoid the stale VP table entry
in the VP profile. This will ensure the proper profile is chosen when
the attach happens after MODE0 recovery is done.

On IPQ5424 platform, during MODE0 recovery, ath12k_ahb_ppeds_free_interrupts()
and ath12k_ahb_ppeds_register_interrupts() are called for freeing and
registering back the DS interrupts. In this sequence, devm_free_irq()
should be called in ath12k_ahb_ppeds_free_interrupts() since the irq
was requested using devm_request_irq(). By not calling the devm_free_irq()
in case of MODE0 recovery, might keep the stale interrupt entry.
When trying to register the IRQ in ath12k_ahb_ppeds_register_interrupts(),
it will fail and leads to a traffic stall in DS mode.

Change-Id: I66eee9625a6b70a031a719fd9bab1254980b413e
Signed-off-by: Tamizh Chelvam Raja <quic_tamizhr@quicinc.com>