UPSTREAM:PROTOCOL:Fix the expiry time of CAC timer
The CAC timeout utilizes the HR timer, which requires the expiry time to be specified in the unit of ktime. The ktime_set() API is currently
used to calculate the ktime from cac_time_ms. This API accepts seconds
and nanoseconds as arguments. Presently, cac_time_ms * NSEC_PER_MSEC is passed to this API as the second argument, which is of the type unsigned long. However, since cac_time_ms * NSEC_PER_MSEC results in a value that can exceed the capacity of an unsigned long variable in a 32-bit machine, an overflow can occur, which can lead to an incorrect calculation of ktime.

To address this issue, replace the ktime_set() API with ms_to_ktime(),
which directly takes milliseconds as input to convert cac_time_ms to ktime.

Fixes: I5251b7bc("UPSTREAM:PROTOCOL: Replace CAC timeout delayed
		 work with HR timer")

Change-Id: I4ffdc120434bad7e9c4f6946af773e59bb6d9067
Signed-off-by: Amith A <quic_amitajit@quicinc.com>