QSDK:CP : Add support for country 00 in mac80211.sh
When the country is set to 00 in the UCI, the output of iw reg get still shows the regulatory information of the previously configured country (US).

As part of the OpenWrt upgrade for ucode, the OpenWrt mac80211.sh replaced the QSDK mac80211.sh. Due to this, the changes for setting the country code for worldwide mode (00) were missed in the OpenWrt mac80211.sh.

To fix this issue, add support for country 00 in the OpenWrt mac80211.sh.

Fixes: I2299484caf0 ("QSDK:CP:mac80211.sh: Add support to bringin features with openwrt upgrade")
Patch-dependency: None
Feature: REGULATORY

Change-Id: I0b591db220e2a440f35ec8f19ab1bb103a9d4932
Signed-off-by: Hariharan Basuthkar <quic_hbasuthk@quicinc.com>