QSDK:DP: wifi: Validate tag in skb->mark before checking MLO assist
Ideally in valid metadata scenario, the skb->mark should start
with the tag as 0xAA; and in other scenarios, it is observed
to be 0.

Looks like the no mark scenario is working fine; whereas issue
with internet traffic is observed only when the metdata is sent
with valid info.
Metadata tag validation was missing at host end.

So added sanity check to validate the tag to be valid before
checking the MLO assist bit(23rd bit in metadata).
This fixes the issue.

Also, check if skb->fast_xmit is set while updating
DP_TX_DESC_FLAG_FAST in tx_desc->flags

Patch-dependency: None
Patch-work: None
Fixes: None

Change-Id: Ie21963a0b59a8280445360d6b412a9eba26b1230
Signed-off-by: Adwait Nayak <quic_adwaitn@quicinc.com>