UPSTREAM:CP wifi:ath12k Set Bridge link on mlo flag during vdev start
Bridge Vdev is created with the same MLD mac-address as of AP vdev
if it under same MLD. When the AP vdev goes down or is removed using
ML reconfiguration, the firmware (FW) will free the ml_bss associated
with the vdev, which the Bridge vdev also shares due to having the
same MLD address.

This results in a FW assert during Bridge peer association, where
accessing the Bridge vdev points to a dangling pointer.

To fix this, set the ATH12K_WMI_FLAG_MLO_BRIDGE_LINK (Bit 14) on the
MLO flags during VDEV_START if it is a Bridge vdev. Also, set the
same MLO flags bit for partner vdev information if it is a Bridge vdev.

Below is the Firmware commit, where ml_bss wont gets freed if bridge vdev
is associated the Normal Vdev during down.
https://qctcollab.qualcomm.com:8443/ui#review:id=3760933

Patch-dependency:
	961-02-wifi-ath12k-n-link-mlo-bridge-vdev-create-start.patc

Change-Id: I8be32b98e2c8efe051d46357b222a4de9ee8694e
Signed-off-by: Vignesh C <quic_vignc@quicinc.com>