UPSTREAM:CP ath12k:wifi Skip peer delete only when link getting down is the primary link
In a high volume station environment, the host is experiencing a
timeout for peer delete WMI because the firmware waits for all
non-primary peer deletions to complete before sending the primary
link peer delete response.

During the peer_delete_all process, peers associated with the link
going down are skipped and included in the
WMI_VDEV_DELETE_ALL_PEER_CMDID command. If the skipped link is
non-primary and the primary link peer delete has already started,
a peer delete response timeout occurs as it takes longer to clear
all peers and send the WMI_VDEV_DELETE_ALL_PEER_CMDID command.

To fix this, skip the peer delete only when the link is primary.
For all non-primary link peer deletions, the firmware will respond
with a peer delete response, and the primary link peer will be
deleted as part of the WMI_VDEV_DELETE_ALL_PEER_CMDID command.

Patch-dependency:
	911-a-wifi-mac80211-Add-wdev-specific-flag-to-indicate-NET.patch
	691-0001-cfg80211-mac80211-add-support-for-channel-switch-wit.patch

Patch-work: none

Change-Id: I559219335ecc1e85ca219d531eb34d551e877a4b
Signed-off-by: Vignesh C <quic_vignc@quicinc.com>