UPSTREAM:CP ath12k:wifi Fix re-addition of STATION vif peer failure during Mode0
During Mode0 recovery for STA vdev, peer re-addition is not being
properly executed because one of the links is not re-added due
to arsta retrieval handling.

To fix this, retrieve the arsta from ahsta based on the passed
link_id and pass the arsta to mac_assign_link_sta. This will map
the arsta in ahsta, ensuring that during further processing,
the peer will be deleted during peer cleanup after the deauth
is sent.

Also, ARRAY_SIZE((ahvif)->link), which calculates the size of
the array at runtime. If there are any discrepancies or changes
in the array size, it could lead to potential errors or unexpected
behavior. Use a predefined constant like ATH12K_NUM_MAX_LINKS
mitigates this risk.

Patch-dependency:
	801-04-wifi-ath12k-Add-support-for-Mode-1-recovery-cache.patch

Change-Id: Ic22b8eaef69684d60f2403bf4b76d0a0802b488e
Signed-off-by: Vignesh C <quic_vignc@quicinc.com>